Change configuration properties and fix store and data entity, converter classes

+++ src/main/java/com/example/data/dto/CompanyDto.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -16,7 +15,7 @@ public class CompanyDto {
 
     private String name;
 
-    private BigDecimal enterpriseFund;
+    private Double enterpriseFund;
 
     private List<WorkerDto> workers;
